EP-818. Moved probability_distribution package under computations.
