+++ tvb_epilepsy/base/computations/probability_distributions/beta_distribution.py
@@ -2,10 +2,10 @@ from collections import OrderedDict
 import numpy as np
 import numpy.random as nr
 import scipy.stats as ss
-from tvb_epilepsy.base.model.statistical_models.probability_distributions import ProbabilityDistributionTypes
+from tvb_epilepsy.base.computations.probability_distributions import ProbabilityDistributionTypes
 from tvb_epilepsy.base.utils.data_structures_utils import isequal_string, make_float
-from tvb_epilepsy.base.model.statistical_models.probability_distributions.continuous_probability_distribution  \
-                                                                                import ContinuousProbabilityDistribution
+from tvb_epilepsy.base.computations.probability_distributions.continuous_probability_distribution \
+    import ContinuousProbabilityDistribution
 
 
 class BetaDistribution(ContinuousProbabilityDistribution):
