+++ tvb_epilepsy/base/computations/probability_distributions/poisson_distribution.py
@@ -1,9 +1,9 @@
 import numpy as np
 import numpy.random as nr
 import scipy.stats as ss
-from tvb_epilepsy.base.model.statistical_models.probability_distributions import ProbabilityDistributionTypes
+from tvb_epilepsy.base.computations.probability_distributions import ProbabilityDistributionTypes
 from tvb_epilepsy.base.utils.data_structures_utils import make_float, make_int, isequal_string
-from tvb_epilepsy.base.model.statistical_models.probability_distributions.discrete_probability_distribution \
+from tvb_epilepsy.base.computations.probability_distributions.discrete_probability_distribution \
     import DiscreteProbabilityDistribution
