+++ tvb_epilepsy/base/computations/probability_distributions/uniform_distribution.py
@@ -2,9 +2,9 @@ import numpy as np
 import scipy.stats as ss
 import numpy.random as nr
 from collections import OrderedDict
-from tvb_epilepsy.base.model.statistical_models.probability_distributions import ProbabilityDistributionTypes
+from tvb_epilepsy.base.computations.probability_distributions import ProbabilityDistributionTypes
 from tvb_epilepsy.base.utils.data_structures_utils import make_float, isequal_string
-from tvb_epilepsy.base.model.statistical_models.probability_distributions.continuous_probability_distribution \
+from tvb_epilepsy.base.computations.probability_distributions.continuous_probability_distribution \
     import ContinuousProbabilityDistribution
 
 DEFAULT_LOW_VALUE = 0.0
